1.67

+++ numpy/01a_np_ndarray_test2.py
@@ -2,8 +2,9 @@
 import numpy as np
 
 # np.empty(shape, dtype=float, order='C') creates an ndarray with random values
-# of the specified shape and dtype,
-arr = np.empty([2, 3], dtype=float)
+# of the specified shape and dtype.
+# shape=[rows, columns] or int, where int is the length of the array.
+arr = np.empty([2, 3], dtype=int)
 print('empty:')
 print(arr)
 print()
